requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
sentence-transformers==3.3.1
joblib==1.3.2
pandas==2.0.2
scikit-learn==1.6.0
google-genai==0.3.0
pytest==8.3.4
httpx==0.28.1
python-multipart==0.0.20